Firebase Analytics new functionality
Moved the code that initializes Firebase to a separate script which is called by the camera when starting the app.

This will allow us to call the Analytics methods in other clases just by importing the libraries
