+++ Analytics.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a86beb9bdc55bb6439deaf2bf71a1da4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
